refactor(tests): create folder unit for tests
