+++ test/unit/test.bufferSave.js
@@ -6,8 +6,8 @@
 
 import test from 'ava'
 import {TestScheduler, ReactiveTest} from 'rx'
-import bufferSave from '../src/bufferSave'
-import {createTestObserver} from '../perf/utils'
+import bufferSave from '../../src/bufferSave'
+import {createTestObserver} from '../../perf/utils'
 const {onNext, onCompleted} = ReactiveTest
 
 test((t) => {
