+++ test/unit/test.contentLoad.js
@@ -7,8 +7,8 @@
 import test from 'ava'
 import { TestScheduler, ReactiveTest } from 'rx'
 const {onNext, onCompleted} = ReactiveTest
-import contentLoad from '../src/contentLoad'
-import { createTestObserver } from '../perf/utils'
+import contentLoad from '../../src/contentLoad'
+import { createTestObserver } from '../../perf/utils'
 
 const noop = function () {}
 test('request', (t) => {
