+++ test/unit/test.initMeta.js
@@ -1,7 +1,7 @@
-import initMeta from '../src/initMeta'
+import initMeta from '../../src/initMeta'
 import test from 'ava'
 import { TestScheduler, ReactiveTest } from 'rx'
-import { createTestObserver } from '../perf/utils'
+import { createTestObserver } from '../../perf/utils'
 const {onNext, onCompleted} = ReactiveTest
 
 test((t) => {
